fin du service web php

+++ base-de-donnees/data/init_sunbay.sql
@@ -9,7 +9,7 @@ CREATE TABLE type_donnee_mesuree(
 
 CREATE TABLE donnee_mesuree(
     id_donnee_mesuree serial PRIMARY KEY,
-    valeur float,
+    valeur numeric(2),
     instant timestamp without time zone,
     id_type_donnee_mesuree integer,
     CONSTRAINT type_donnee_mesuree_mesure_fk
